move files into handler convert messageUpdate to TS
