+++ src/events/messageUpdate.ts
@@ -1,165 +1,44 @@
-// const configFileName = process.env.NIRA_DEV ? 'config.dev.json' : 'config.json'
-// const { allowlists, members, emojis } = require(`../${configFileName}`)
+import { Message } from 'discord.js';
+import { CommandoClient } from 'discord.js-commando';
 
-// // Find specific emojis in a newMessage
-// function matchEmojis(find_emojis, message_content) {
-//     const emoji_regexp = /<a?:\w+:\d+>/g
-//     const matches = [...message_content.matchAll(emoji_regexp)]
-//     let matched_emojis = []
-//     matches.forEach((match) => {
-//         if (find_emojis.includes(match[0])) {
-//             matched_emojis.push(match[0])
-//             if (match[0] === emojis.owie) {
-//                 matched_emojis.push(emojis.cursed)
-//             }
-//         }
-//     })
-//     return matched_emojis
-// }
+import {
+  handleCountingEditedMessage,
+  handleDeathOfNira,
+  handleNiraWave,
+  handleNoUMessage,
+  handlePaladinMessage,
+  handlePoyoMessage,
+  handleCheckNiraMojis,
+  handleNiraMessageReactions,
+  handleNiraMojisInChannels,
+  handlePatPatRole,
+  handleTwoWordStoryMessage,
+} from '../handler';
 
-// module.exports = async (client, oldMessage, newMessage) => {
-//     // Ignore event if content has not changed
-//     if (newMessage.content === oldMessage.content) {
-//         return
-//     }
+export default async function (client: CommandoClient, oldMessage: Message, newMessage: Message) {
+  if (newMessage.content === oldMessage.content) {
+    return;
+  }
 
-//     // Check Edited Messages
-//     async function userReactions(newMessage) {
-//         {
-//             const userReactions = newMessage.reactions.cache.filter((reaction) =>
-//                 reaction.users.cache.has(client.user.id),
-//             )
-//             for (const reaction of userReactions.values()) {
-//                 await reaction.users.remove(client.user.id)
-//             }
-//         }
-//     }
-//     await userReactions(newMessage)
+  await handleNiraMessageReactions(newMessage, client);
 
-//     // Check for NiraMojis in their channels
-//     if (allowlists.disgustchannels.includes(newMessage.channel.id)) {
-//         if (![emojis.disgust].includes(newMessage.content)) {
-//             return newMessage.delete()
-//         }
-//     } else if (allowlists.starechannels.includes(newMessage.channel.id)) {
-//         if (![emojis.stare].includes(newMessage.content)) {
-//             return newMessage.delete()
-//         }
-//     } else if (allowlists.owiechannels.includes(newMessage.channel.id)) {
-//         if (![emojis.owie].includes(newMessage.content)) {
-//             return newMessage.delete()
-//         }
-//     }
+  handleNiraMojisInChannels(newMessage);
 
-//     // Check for NiraMojis everywhere
-//     if (
-//         newMessage.content.includes(emojis.disgust) ||
-//         newMessage.content.includes(emojis.stare) ||
-//         newMessage.content.includes(emojis.owie)
-//     ) {
-//         const find_emojis = [emojis.disgust, emojis.stare, emojis.owie]
-//         let matched_emojis = matchEmojis(find_emojis, newMessage.content)
+  handleCheckNiraMojis(newMessage);
 
-//         matched_emojis.forEach((e) => newMessage.react(e))
-//     }
+  handlePatPatRole(newMessage);
 
-//     // PatPat Role
-//     // The member attribute is undefined on some messages so check if it's defined first
-//     if (newMessage.member && newMessage.member.roles.cache.get('765347466169024512')) {
-//         if (newMessage.content.toLowerCase().includes('patpat', emojis.patpat)) {
-//             newMessage.react('761487227921367051')
-//         }
-//     }
+  handleNiraWave(newMessage, client);
 
-//     // Nira Wave
-//     if (newMessage.mentions.users.has(client.user.id)) {
-//         newMessage.react('742394597174673458')
-//     }
+  handleCountingEditedMessage(newMessage);
 
-//     // Counting
-//     if (newMessage.channel.id === '758541031498317835') {
-//         newMessage.delete()
-//     }
+  handleTwoWordStoryMessage(newMessage);
 
-//     // 2-Word Story Channel
-//     if (newMessage.channel.id === '776229267998375946') {
-//         const args = newMessage.content.trim().split(/ +/g)
-//         if (!args[1] || args[2]) {
-//             newMessage.delete()
-//         }
-//     }
+  handleDeathOfNira(newMessage);
 
-//     // Death of Nira
-//     const testingNira = '764990952510717973'
-//     const niraWave = emojis.wave.replace(/\D/g, '')
-//     if (newMessage.mentions.users.has(testingNira)) {
-//         const filter = (reaction, user) => user.id === testingNira && reaction.emoji.id === niraWave
+  handlePoyoMessage(newMessage);
 
-//         newMessage.awaitReactions(filter, { max: 1, time: 3500 }).then((collected) => {
-//             if (!collected.size) {
-//                 newMessage.react('756582453824454727')
-//             }
-//         })
-//     }
+  handlePaladinMessage(newMessage);
 
-//     // Responses
-//     if (newMessage.content.toLowerCase().includes('poyo')) {
-//         if (newMessage.content.toLowerCase() === ('poyo' || 'poyo!') || Math.random() < 1 / 2) {
-//             newMessage.channel.send('Poyo!')
-//         }
-//     }
-//     const winemoji = '802352890571653170'
-//     var paladinResponses = ['decant', 'grand paladin', `<:1945grandpaladin:${winemoji}>`]
-//     let isPaladin = paladinResponses.some((word) => newMessage.content.toLowerCase().includes(word.toLowerCase()))
-//     if (isPaladin) {
-//         if (newMessage.content === 'Decant.') {
-//             newMessage
-//                 .react('🇩')
-//                 .then(() => newMessage.react('🇪'))
-//                 .then(() => newMessage.react('🇨'))
-//                 .then(() => newMessage.react('🇦'))
-//                 .then(() => newMessage.react('🇳'))
-//                 .then(() => newMessage.react('🇹'))
-//         } else {
-//             newMessage.react(winemoji)
-//             if (Math.random() < 1 / 10) {
-//                 const embed = new MessageEmbed()
-//                     .setDescription(
-//                         "It is perfection. Irreplaceable. You don't drink the **1945 Grand Paladin** anymore than you'd write a shopping list on the Mona Lisa!",
-//                     )
-//                     .setColor(15849719)
-//                 newMessage.channel.send(embed)
-//             }
-//         }
-//     }
-//     var noUResponses = [
-//         'no u',
-//         "yesn't men't",
-//         'nay thee',
-//         '[Rn] 5f¹⁴7s² × [Rn] 5f³6d¹7s²',
-//         'n-nyo u~wu',
-//         'Nobelium Uranium',
-//         'non tu',
-//         'no vos',
-//         '102 + 92',
-//         '`6e 6f 20 75`',
-//         '🇳 🇴  🇺',
-//         'ノユ',
-//         '∩O ∪',
-//         '∩∅ ∪',
-//         '`01101110 01101111 00100000 01110101`',
-//         '`-. --- / ..-`',
-//         '`110 111 32 117`',
-//         '`&#110;&#111;&#32;&#117;`',
-//         'ⁿᵒ ᵘ',
-//     ]
-//     let isNoU = noUResponses.some((word) => newMessage.content.toLowerCase() === word.toLowerCase())
-//     if (isNoU && (Math.random() < 1 / 3 || members.noutimesinfinity.includes(newMessage.author.id))) {
-//         const response = noUResponses[Math.floor(Math.random() * noUResponses.length)]
-//         newMessage.channel.send(response)
-//     }
-// }
-
-export default function () {
-  console.log('hi');
+  handleNoUMessage(newMessage);
 }
